Updated Secret Encoding Mistake

+++ routes/user.ts
@@ -8,7 +8,7 @@ import { checkAuth } from "../util.ts";
 
 const router = express.Router();
 
-//GET ALL USERS
+//doc
 router.get("/", async (req: Request, res: Response) => {
     try {
         if(req.headers.authorization==null || req.headers.authorization!==`${Deno.env.get("ADMIN_TOKEN")}`){
@@ -17,79 +17,91 @@ router.get("/", async (req: Request, res: Response) => {
         const users: UserType[] = await User.find().select("-password -__v -_id");
         res.status(200).json(users);
     } catch (err: Error | any) {
-        res.status(500).json({ error: err.message });
+        res.status(500).json({ error: "Internal Server Error" });
     }
 });
 
-//GET ONE USER BY ID
+//doc
 router.get("/:userid", async (req: Request, res: Response) => {
     try {
         const userid = req.params.userid
+        if (!userid) {
+            return res.status(400).json({ error: "Missing params" });
+        }
         const isAuth = await checkAuth(userid,req.cookies.bearer)
         if(!isAuth){
-           return res.status(400).json({ error: "Not yourself or not admin" });
+           return res.status(401).json({ error: "Unauthorized" });
         }
         const user: UserType | null = await User.findOne({ userid }).select("-password -__v -_id");
         if(!user){
-            return res.status(404).json({ error: "User not found" });
+            return res.status(404).json({ error: "Not found" });
         }
         res.status(200).json(user);
     } catch (err: Error | any) {
-        res.status(500).json({ error: "Error interno servidor" });
+        res.status(500).json({ error: "Internal Server Error" });
     }
 });
 
-//GET TICKETS OF A USER BY USERID
+//doc
 router.get("/tickets/:userid", async (req: Request, res: Response) => {
     try {
         const userid = req.params.userid
+        if (!userid) {
+            return res.status(400).json({ error: "Missing params" });
+        }
         const isAuth = await checkAuth(userid,req.cookies.bearer)
         if(!isAuth){
-           return res.status(400).json({ error: "Not yourself or not admin" });
+           return res.status(401).json({ error: "Unauthorized" });
         }
         const tickets: TicketType[] | null = await Ticket.find({ userid }).select("-__v -_id");
         if(!tickets){
-            return res.status(404).json({ error: "User has no tickets" });
+            return res.status(404).json({ error: "Not found" });
         }
         res.status(200).json(tickets);
     } catch (err: Error | any) {
-        res.status(500).json({ error: "Error interno servidor" });
+        res.status(500).json({ error: "Internal Server Error" });
     }
 });
 
-//DELETE A USER BY ID
+//doc
 router.delete("/:userid", async (req: Request, res: Response) => {
     try {
         const userid = req.params.userid
+        if (!userid) {
+            return res.status(400).json({ error: "Missing params" });
+        }
         const isAuth = await checkAuth(userid,req.cookies.bearer)
         if(!isAuth){
-           return res.status(400).json({ error: "Not yourself or not admin" });
+           return res.status(401).json({ error: "Unauthorized" });
         }
         const deletedUser = await User.deleteOne({ userid });
         if(deletedUser){
-            return res.status(200).json({ message: "User Deleted Succesfully" });
+            return res.status(200).json({ message: "User Deleted" });
         }
-        res.status(401).json({error: "Error Deleting User"});
+        res.status(404).json({error: "Not found"});
         } catch (err: Error | any) {
-        res.status(500).json({ error: "Error interno servidor" });
+        res.status(500).json({ error: "Internal Server Error" });
     }
 });
 
-//UPDATE A USER BY ID
+//doc
 router.put("/", async (req: Request, res: Response) => {
     try {
             const userid = req.body.userid
+            if (!userid) {
+            return res.status(400).json({ error: "Missing params" });
+        }
             const isAuth = await checkAuth(userid,req.cookies.bearer)
             if(!isAuth){
-            return res.status(400).json({ error: "Not yourself or not admin" });
+            return res.status(401).json({ error: "Unauthorized" });
             }
             if(req.body.userid==null || req.body.name==null || req.body.email==null|| req.body.password==null){
-                return res.status(400).json({ error: "Faltan datos obligatorios" });
+                return res.status(400).json({ error: "Missing params" });
             }
             const hashedPassword = await bcrypt.hash(req.body.password,10);
             const useractual = await User.findOne({ userid: req.body.userid });
             if(!useractual){
-                return res.status(404).json({ error: "User not found" });
+                return res.status(404).json({ error: "Not Found" });
             }
             const coins = useractual ? useractual.coins : "0";
             await User.updateOne(
@@ -98,7 +110,7 @@ router.put("/", async (req: Request, res: Response) => {
             );
             res.status(200).json({message:"User Updated Succesfully"});
         } catch (err: Error | any) {
-            res.status(500).json({ error: "Error interno servidor" });
+            res.status(500).json({ error: "Internal Server Error" });
         }
 });
